Migrate plugin tabs to full paths
* #releases => /plugin.name/releases
* #issues => /plugin.name/issues
* #dependencies => /plugin.name/dependencies
* add version_${version.id) to releases
* Add redirects to old anchors to full urls
